Print semicolon as part of statement in javascript Printer
This is necesary for a subsequent change, where we pre-print
statements into byte buffers. When we later assemble the statement
into the surrounding block, we do not have type information
anymore (and hence cannot judge whether a semicolon is necessary).

Note that this now prints a semicolon after the last statement in a
block (which we didn't previously). This does increase fastopt
size (somewhat artificially), but more closely corresponds:

- to what a human would write
- the ECMAScript spec (e.g. https://tc39.es/ecma262/#sec-expression-statement)